+++ src/components/datepicker/constant.js
@@ -10,7 +10,6 @@ export const containerClass = `${selectorClass}-container`;
 export const calendarIcon = <Icon type="calendar" className={`${selectorClass}-inp-icon`}></Icon>;
 
 export const monthArr = ['一月', '二月', '三月', '四月', '五月', '六月', '七月', '八月', '九月', '十月', '十一月', '十二月'];
-export const miniWeek = ['日', '一', '二', '三', '四', '五', '六'];
 
 export const enumObj = {
 	LEFT: 'LEFT',
@@ -26,8 +25,5 @@ export const enumObj = {
 
 	AUTO: 'AUTO',
 	UP: 'UP',
-	DOWN: 'DOWN',
-
-	START: 'START',
-	END: 'END'
+	DOWN: 'DOWN'
 };
